% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landmark_monster.R
\name{create.landmark.set}
\alias{create.landmark.set}
\title{Create landmark monster}
\usage{
create.landmark.set(n.landmarks = 5, reflect = FALSE, base.seed = NA)
}
\arguments{
\item{n.landmarks}{number of landmarks}

\item{reflect}{reflect}

\item{base.seed}{seed}
}
\description{
Create landmark monster
}
