% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_landmark_set.R
\name{create.test.landmark.set}
\alias{create.test.landmark.set}
\title{Create test landmark set}
\usage{
create.test.landmark.set(n.landmarks = 5, reflect = FALSE,
  n.monsters = 10, sd = 0.1, base.seed = NA, random.seed = NA)
}
\arguments{
\item{n.landmarks}{number of landmarks}

\item{reflect}{reflect}

\item{n.monsters}{number of monsters}

\item{sd}{randomization standard deviation}

\item{base.seed}{base monster seed}

\item{random.seed}{randomization seed}
}
\description{
Create test landmark set
}

