% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform_set.R
\name{deform.set}
\alias{deform.set}
\title{Deform landmark set}
\usage{
deform.set(data, a.interval, theta.interval, noise.a.sd = 0.01,
  noise.theta.sd = 0.01, deform.set.seed = NA)
}
\arguments{
\item{data}{data}

\item{a.interval}{values of a}

\item{theta.interval}{values of theta}

\item{noise.a.sd}{standard deviation of a}

\item{noise.theta.sd}{standard deviation of theta}

\item{deform.set.seed}{deformation seed}
}
\description{
Deform landmark set
}

