% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform_with_noise.R
\name{deform.with.noise}
\alias{deform.with.noise}
\title{Deform landmark monster}
\usage{
deform.with.noise(data, a, theta, noise.a.sd = 0.01, noise.theta.sd = 0.01,
  deform.seed = NA)
}
\arguments{
\item{data}{landamrk monster}

\item{a}{a}

\item{theta}{theta}

\item{noise.a.sd}{standard deviation of a}

\item{noise.theta.sd}{standard deviation of theta}

\item{deform.seed}{deformation seed}
}
\description{
Deform landmark monster
}

