% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_monster.R
\name{randomize.monster}
\alias{randomize.monster}
\title{Create random set of landmark monsters}
\usage{
randomize.monster(base.monster, n.monsters, sd = 0.1, reflect = FALSE,
  random.seed = NA)
}
\arguments{
\item{base.monster}{initial monster}

\item{n.monsters}{number of monsters}

\item{sd}{standard deviation of randomization}

\item{reflect}{reflect}

\item{random.seed}{seed}
}
\value{
Set of landmark monsters
}
\description{
Create random set of landmark monsters
}

